upload plugin tests
