+++ core/tests/mocks/.gitignore
@@ -1,0 +1,3 @@
+
+# zip files should be created just in time for tests and deleted afterwards
+*.zip
